@@ -2,7 +2,7 @@
   "license": "Apache-2.0",
   "private": true,
   "engines": {
-    "node": ">=21.0.0"
+    "node": ">=22.0.0"
   },
   "workspaces": [
     "@shared/api",
