@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useRef, useState } from "react";
 import { useSelector } from "react-redux";
-import { Link, Redirect } from "react-router-dom";
+import { Link, Navigate } from "react-router-dom";
 import { Formik, Form, Field, FieldProps } from "formik";
 import debounce from "lodash/debounce";
 import { useTranslation } from "react-i18next";
@@ -216,7 +216,7 @@ export const SearchAsset = () => {
   }, [assetRows, networkDetails]);
 
   if (isCustomNetwork(networkDetails)) {
-    return <Redirect to={ROUTES.addAsset} />;
+    return <Navigate to={ROUTES.addAsset} />;
   }
 
   return (
