@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { AppDispatch } from "popup/App";
 import {
@@ -37,8 +37,8 @@ export const SelectAssetRows = ({ assetRows }: SelectAssetRowsProps) => {
     soroswapTokens,
     transactionData,
   } = useSelector(transactionSubmissionSelector);
-  const dispatch: AppDispatch = useDispatch();
-  const history = useHistory();
+  const dispatch = useDispatch<AppDispatch>();
+  const navigate = useNavigate();
   const isSoroswapEnabled = useIsSoroswapEnabled();
   const isSwap = useIsSwap();
 
@@ -109,11 +109,11 @@ export const SelectAssetRows = ({ assetRows }: SelectAssetRowsProps) => {
                   if (assetSelect.isSource) {
                     dispatch(saveAsset(canonical));
                     dispatch(saveIsToken(isContract));
-                    history.goBack();
+                    navigate(-1);
                   } else {
                     dispatch(saveDestinationAsset(canonical));
                     dispatch(saveDestinationIcon(icon));
-                    history.goBack();
+                    navigate(-1);
                   }
                   dispatch(saveIsSoroswap(isSoroswap));
                 }}
