@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { Redirect } from "react-router-dom";
+import { Navigate } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { Button, Text, Icon } from "@stellar/design-system";
 
@@ -15,6 +15,7 @@ import {
   confirmMnemonicPhrase,
 } from "popup/ducks/accountServices";
 import { View } from "popup/basics/layout/View";
+import { AppDispatch } from "popup/App";
 
 import "./styles.scss";
 
@@ -29,7 +30,7 @@ export const MnemonicPhrase = ({
   const applicationState = useSelector(applicationStateSelector);
   const [isConfirmed, setIsConfirmed] = useState(false);
   const [isModalShowing, setIsModalShowing] = useState(true);
-  const dispatch = useDispatch();
+  const dispatch = useDispatch<AppDispatch>();
 
   const handleSkip = () => {
     // confirm the mnemonic phrase for the user
@@ -38,7 +39,7 @@ export const MnemonicPhrase = ({
   };
 
   if (applicationState === APPLICATION_STATE.MNEMONIC_PHRASE_CONFIRMED) {
-    return <Redirect to={ROUTES.mnemonicPhraseConfirmed} />;
+    return <Navigate to={ROUTES.mnemonicPhraseConfirmed} />;
   }
 
   if (mnemonicPhrase) {
