@@ -21,6 +21,7 @@ import { AccountHeaderModal } from "popup/components/account/AccountHeaderModal"
 import { NetworkIcon } from "popup/components/manageNetwork/NetworkIcon";
 
 import "./styles.scss";
+import { AppDispatch } from "popup/App";
 
 interface AccountHeaderProps {
   allAccounts: Account[];
@@ -36,7 +37,7 @@ export const AccountHeader = ({
   setLoading,
 }: AccountHeaderProps) => {
   const { t } = useTranslation();
-  const dispatch = useDispatch();
+  const dispatch = useDispatch<AppDispatch>();
   const networkDetails = useSelector(settingsNetworkDetailsSelector);
   const networksList = useSelector(settingsNetworksListSelector);
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
@@ -99,12 +100,10 @@ export const AccountHeader = ({
             <li className="AccountHeader__account-list-item">
               <Link
                 className="AccountHeader__account-list-item__link"
-                to={{
-                  pathname: ROUTES.addAccount,
-                  state: {
-                    header: t("Create a new Stellar address"),
-                    cta: t("Add address"),
-                  },
+                to={ROUTES.addAccount}
+                state={{
+                  header: t("Create a new Stellar address"),
+                  cta: t("Add address"),
                 }}
               >
                 <div className="AccountHeader__account-list-item__row">
