@@ -40,7 +40,7 @@ export const ModifyAssetList = ({
   const location = useLocation();
   const params = new URLSearchParams(location.search);
   const assetListUrl = params.get("asset-list-url");
-  const dispatch: AppDispatch = useDispatch();
+  const dispatch = useDispatch<AppDispatch>();
   const [fetchErrorString, setFetchErrorString] = useState("");
   const [submitErrorString, setSubmitErrorString] = useState("");
   const [assetListInfo, setAssetListInfo] = useState({} as AssetsListsData);
